backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
anthropic==0.28.0
PyGithub==2.3.0
PyJWT==2.8.0
cryptography==42.0.8
httpx==0.27.0
unidiff==0.7.5
pyyaml==6.0.1
pip-audit==2.7.3
google-cloud-firestore==2.16.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.10.0
pytest==8.2.0
pytest-asyncio==0.23.7
